+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Unreleased
+##  v1.1.0
+### Added
+- `GET /sapi/v1/fiat/orders` to query user fiat deposit and withdraw history
+- `GET /sapi/v1/fiat/payments` to query user fiat payments history
 
+
+## v1.0.0
 ### Added
 - First release, please find details from `README.md`